[CI] Use actions/checkout v4
Part of OpenAssetIO/OpenAssetIO#1351. Now that the `openassetio-build`
container can support nodejs 20 (because glibc is new enough) we can
update `actions/checkout`.

This was an attempt to fix an error in GitHub CI

```
fatal: detected dubious ownership in repository at '/__w/OpenAssetIO-Test-CMake/OpenAssetIO-Test-CMake'
```

But it didn't work. See subsequent commit for workaround.

Signed-off-by: David Feltell <david.feltell@foundry.com>